+++ apps/frontend/src/config/navigationFlowchart.schema.ts
