+++ apps/frontend/src/config/navigationFlowchart.ts
