+++ apps/frontend/src/scripts/database/seed-navigation-flowcharts.ts
@@ -6,7 +6,7 @@ import type {
   NavigationFlowchartConfig,
   NavigationFlowchartOption,
   NavigationFlowchartStep,
-} from "@/config/navigation-flowchart";
+} from "@/config/navigationFlowchart";
 import * as schema from "@/db/schema";
 import { NAVIGATION_FLOWCHART_STATUS } from "@/db/schema";
